.ipynb_checkpoints/requirements-checkpoint.txt
livekit-agents
llama-index
python-dotenv
openai
elevenlabs
